+++ test/dummy_app/app/models/parent_without_children.rb
@@ -1,0 +1,5 @@
+class ParentWithoutChildren < ApplicationRecord
+  include ActiveSnapshot
+
+  self.table_name = "posts"
+end